Fix regex key mapping example
The example given is an incomplete copy of a test case
tests/files/success/16s.yaml
but because the schema is incomplete and the data section is
complete, this breaks the test case. Consequently the example
code did not work.

I have fixed this by removing the leftover entry in Data so the
example Data is compliant with the existing Schema.