Do not fix dependencies version
pyyaml released few days ago a new minor release, namely 3.12. As Pykwalify explicitely relies on 3.11, this means that many projects could not benefit from the new version. Moreover, as pip isn't bundled with a dependency contraints solver, this means that any requirements.txt that specify pyyaml and pykwalify (in this order) will not be satisfied (because pyyaml will be first installed in 3.12, then pykwalify will require 3.11, which is naturally in conflict with 3.12).