core: allow custom functions to return exceptions
Currently, custom functions can signal a schema failure only by raising
an exception. This means that instead of seeing a nice readout of all
schema failures in a list (as you normally would), you see an immediate
failure of just one exception with a stack trace. In addition, it is up
to the error message how much detail to provide, so sometimes you don't
have enough information to easily fix the issue.

Improve this by allowing custom functions to return an object that can
be turned into a string (can call str() or unicode() on it) instead of
raising an exception. The returned string will be collected in the list
of errors and displayed at the end, along with path information, in the
same way as the standard errors are displayed.